Refactor tracking of overall package selection state
This fixes two issues with the anySelected and allSelected properties:
- Selecting an update broke the binding of anySelected, breaking the
  "unchecked" state forever
- They were bindings, which got reevaluated any time updatesModel changed,
  particularly in the populateModel loop, which is incredibly expensive

Move the tracking outside of bindings and take care of updating manually.
This can be done in a single iteration now, merging both functions.

The result is that changes to the model are now much faster and the checkbox
state is now correct.