Don't default to showing that the check for updates failed
On startup, it doesn't do any PackageKit transaction if necessary
(intentionally disabled in 4150bd6264d209b300598abea14fd9fce63c2f6e), which
means that m_lastCheckSuccessful stays at false. This results in the default
message on startup being that "Checking for updates failed", which is
fortunately wrong.

Fix this by introducing a separate state for just this.